refactor(backup): streamline mailbox progress tracking by integrating a dedicated progress adapter and enhance CSV parsing for mailbox usage
