+++ src/cli/commands/mailboxes.command.ts
@@ -5,6 +5,7 @@ import { ATLAS_CONFIG_TOKEN } from '@/utils/config';
 import type { MailboxDiscoveryService, TenantMailbox } from '@/ports/mailbox/discovery.port';
 import { MAILBOX_DISCOVERY_TOKEN } from '@/ports/tokens/outgoing.tokens';
 import { logger } from '@/utils/logger';
+import { format_bytes, pad_cell, truncate_cell } from '@/cli/command-formatters';
 
 type ContainerFactory = () => Container;
 
@@ -32,9 +33,9 @@ async function execute_mailboxes(container: Container, options: MailboxesOptions
   logger.info(`Tenant: ${tenant_id}`);
 
   const discovery = container.get<MailboxDiscoveryService>(MAILBOX_DISCOVERY_TOKEN);
-  const mailboxes = await discovery.list_tenant_mailboxes(tenant_id, {
-    licensed_only: options.licensedOnly,
-  });
+  const discovery_options =
+    options.licensedOnly === undefined ? undefined : { licensed_only: options.licensedOnly };
+  const mailboxes = await discovery.list_tenant_mailboxes(tenant_id, discovery_options);
 
   if (mailboxes.length === 0) {
     logger.warn(
@@ -60,11 +61,11 @@ function print_mailbox_table(mailboxes: TenantMailbox[]): void {
   const has_sizes = mailboxes.some((m) => m.mailbox_size_bytes !== undefined);
   const header =
     '  ' +
-    pad('Mail', 38) +
-    pad('Display Name', 24) +
-    pad('Exchange', 10) +
-    pad('Status', 10) +
-    (has_sizes ? pad('Size', 12) : '') +
+    pad_cell('Mail', 38) +
+    pad_cell('Display Name', 24) +
+    pad_cell('Exchange', 10) +
+    pad_cell('Status', 10) +
+    (has_sizes ? pad_cell('Size', 12) : '') +
     'Created';
   console.log(header);
   console.log('  ' + '-'.repeat(header.length - 2));
@@ -73,13 +74,13 @@ function print_mailbox_table(mailboxes: TenantMailbox[]): void {
     const license_flag = m.has_exchange_license ? 'Yes' : 'No';
     const status = m.exchange_plan_status ?? '--';
     const created = m.created_at ? m.created_at.toISOString().slice(0, 10) : '--';
-    const size = has_sizes ? pad(format_size(m.mailbox_size_bytes), 12) : '';
+    const size = has_sizes ? pad_cell(format_size(m.mailbox_size_bytes), 12) : '';
     console.log(
       '  ' +
-        pad(truncate(m.mail, 36), 38) +
-        pad(truncate(m.display_name, 22), 24) +
-        pad(license_flag, 10) +
-        pad(status, 10) +
+        pad_cell(truncate_cell(m.mail, 36), 38) +
+        pad_cell(truncate_cell(m.display_name, 22), 24) +
+        pad_cell(license_flag, 10) +
+        pad_cell(status, 10) +
         size +
         created,
     );
@@ -88,16 +89,5 @@ function print_mailbox_table(mailboxes: TenantMailbox[]): void {
 
 function format_size(bytes?: number): string {
   if (bytes === undefined) return '--';
-  if (bytes < 1024) return `${bytes} B`;
-  if (bytes < 1024 * 1024) return `${(bytes / 1024).toFixed(1)} KB`;
-  if (bytes < 1024 * 1024 * 1024) return `${(bytes / (1024 * 1024)).toFixed(1)} MB`;
-  return `${(bytes / (1024 * 1024 * 1024)).toFixed(2)} GB`;
-}
-
-function pad(str: string, width: number): string {
-  return str.padEnd(width);
-}
-
-function truncate(str: string, max: number): string {
-  return str.length > max ? str.slice(0, max - 1) + '~' : str;
+  return format_bytes(bytes, 2);
 }
